requirements.txt
discord.py>=2.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
supabase>=2.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
httpx>=0.25.0

